modify to 141

+++ coding-study/baekjoon/bronze/Num_24265.cpp
@@ -2,7 +2,7 @@
 
 int main(void){
     
-    int N;
+    long long N;
     std::cin >> N;
 
     std::cout << (N * (N-1)) / 2 << "\n" << "2";
